MyCloud/requirements.txt
boto3
azure-common
azure-mgmt-compute
azure-mgmt-managementgroups
azure-mgmt-network
azure-mgmt-resource
haikunator
